requirements.txt
# Ranger Lead Agent - Dependencies

# OpenAI Agents SDK (includes openai SDK)
openai-agents>=0.1.0

# HTTP client for custom API calls
httpx>=0.27.0

# Environment management
python-dotenv>=1.0.0

# Data validation (used by Agents SDK for output_type)
pydantic>=2.0.0

# Optional: Excel support
# pandas>=2.0.0
# openpyxl>=3.1.0

# Web interface
streamlit>=1.30.0
